chore: further improvements regarding the comments & discussions Added new endpoint structures for comments and discussions (regarding only articles for now). Added database utilities for the comments, discussions and their intermediary table discussions_comments.

+++ src/handlers/articles.go
@@ -256,6 +256,10 @@ func DeleteArticle(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
+	if db.DeleteArticlesCommentByArticleId(body.Id).Handle(w, r) {
+		return
+	}
+
 	if db.DeleteArticleById(body.Id).Handle(w, r) {
 		return
 	}
